+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add prometheus remote write ingress boilerplate.
 - Add prometheus agent remote write secret boilerplate.
 - Generate prometheus remote write agent secret and config.
+- Configure prometheus remote write agent ingress.
 
 ## [4.6.2] - 2022-09-13